# docs/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bank REST API",
    "description": "Это API предоставляет эндпоинты для управления банковскими картами.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/user/cards/{cardId}/request-block": {
      "post": {
        "tags": [
          "User Cards"
        ],
        "summary": "Запросить блокировку карты",
        "operationId": "requestBlock",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/user/cards/transfer": {
      "post": {
        "tags": [
          "User Cards"
        ],
        "summary": "Перевод между картами пользователя",
        "operationId": "transferBetweenCards",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
...
    "/api/user/cards/{cardId}/get-decrypted": {
      "get": {
        "tags": [
          "User Cards"
        ],
        "summary": "Получить расшифрованную карту по её ID",
        "operationId": "getDecryptedUserCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cards/{cardId}/balance": {
      "get": {
        "tags": [
          "User Cards"
        ],
        "summary": "Получить баланс карты",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cards/get": {
      "get": {
        "tags": [
          "User Cards"
        ],
        "summary": "Получить замаскированные карты пользователя",
        "operationId": "getUserCards",
        "parameters": [
          {
            "name": "arg0",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "arg1",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardFilterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cards/get-decrypted": {
      "get": {
        "tags": [
          "User Cards"
        ],
        "summary": "Получить расшифрованные карты пользователя",
        "operationId": "getDecryptedUserCards",
        "parameters": [
          {
            "name": "arg0",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferRequest": {
        "required": [
          "amount",
          "fromCardId",
          "toCardId"
        ],
        "type": "object",
        "properties": {
          "fromCardId": {
            "type": "integer",
            "format": "int64"
          },
          "toCardId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "minimum": 0.01,
            "exclusiveMinimum": false,
            "type": "number"
          }
        }
      },
...
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          },
          "maskedNumber": {
            "type": "string"
          },
          "decryptedNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED",
              "PENDING_BLOCK"
            ]
          },
          "balance": {
            "type": "number"
          },
          "expiry": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CardFilterRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED",
              "PENDING_BLOCK"
            ]
          },
          "minBalance": {
            "type": "number"
          },
          "maxBalance": {
            "type": "number"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}